+++ CHANGELOG.md
@@ -6,15 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-08-22
+
+### Changed
+- New class tree comparison algorithm
+- New HTML report template
+- Gradle task name to `shrinkometer`
+
 ## [0.2.1] - 2020-08-20
 
 ### Fixed
-- 'Unable to find method' error by downgrading to Kotlin 1.3.72
+- `Unable to find method` error by downgrading to Kotlin 1.3.72
 
 ## [0.2.0] - 2020-08-20
 
 Initial release
 
-[Unreleased]: https://github.com/Erdenian/shrinkometer/compare/0.2.1...develop
+[Unreleased]: https://github.com/Erdenian/shrinkometer/compare/0.3.0...develop
+[0.3.0]: https://github.com/Erdenian/shrinkometer/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/Erdenian/shrinkometer/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/Erdenian/shrinkometer/releases/tag/0.2.0